==> Uber Pickups/uber-trip-data/step0_formatting_taxizones.ipynb <==
import pandas as pd
nybb=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/taxi_zones.csv")
nybb.head()

nybb['the_geom'][0][16:40]

nybb['coordinates']=nybb['the_geom'].apply(lambda x: x[16:42])
nybb['coordinates'].str.split(' ').to_list()

nybbcoordinates=pd.DataFrame(nybb['coordinates'].str.split(' ',1).tolist(),
                          columns = ['Lon','Lat'])

taxi_zone_coordinates=pd.concat([nybb,nybbcoordinates],axis=1)
taxi_zone_coordinates.head()

taxi_zone_coordinates=taxi_zone_coordinates.drop(['OBJECTID','Shape_Leng','the_geom','Shape_Area','coordinates'],axis=1)

taxi_zone_coordinates['Lat']=taxi_zone_coordinates['Lat'].apply(lambda x:round(float(x),2))
taxi_zone_coordinates['Lon']=taxi_zone_coordinates['Lon'].apply(lambda x:round(float(x),2))
print(taxi_zone_coordinates)

file_name = '/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_taxizones.csv'
taxi_zone_coordinates.to_csv(file_name)
print('File is written to CSV File successfully.')
