==> Uber Pickups/uber-trip-data/step1_formatting_janjune15.ipynb <==
import pandas as pd
janjune15=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-janjune-15.csv")
janjune15.head()

split_datetime=pd.DataFrame(janjune15['Pickup_date'].str.split(' ').tolist(),
                         columns = ['Date','Time'])

split_datetime

#Formatter la date dans le format JJ/MM/AAAA au lieu de AAAA/MM/JJ 
from datetime import datetime
d=pd.to_datetime(split_datetime['Date'],format='%Y-%m-%d')
split_datetime['Date']=d.dt.strftime("%d/%m/%y")
split_datetime.head()

data15=pd.concat([janjune15,split_datetime],axis=1)
data15.shape

## Converger les resultats avec Taxi Zone pour obtenir le détail de l'historique data janjune15

taxi_zone_coordinates=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_taxizones.csv",index_col=[0])
taxi_zone_coordinates=taxi_zone_coordinates.rename(columns={"LocationID": "locationID","zone":"Zone","borough":"Borough"})
taxi_zone_coordinates

result15=data15.merge(taxi_zone_coordinates, on='locationID', how='left')
result15.head()

result15=result15.rename(columns={"Affiliated_base_num": "Base"}).dropna()

result15.head()

mycolumns=['Date','Time','locationID','Base','Lon','Lat','Borough','Zone']
result15=result15[mycolumns]
result15.head()

file_name = '/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_janjune15.csv'
result15.to_csv(file_name)
print('File is written to CSV File successfully.')
