==> Uber Pickups/uber-trip-data/step2_formatting_aprsep14.ipynb <==
import pandas as pd
apr14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-apr14.csv")
may14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-may14.csv")
jun14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-may14.csv")
jul14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-may14.csv")
aug14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-may14.csv")
sep14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-may14.csv")

import pandas as pd
list=[]
month14=['apr14','may14','jun14','jul14','aug14','sep14']
for month in range(len(month14)):
    link="/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/uber-trip-data/uber-raw-data-{}.csv".format(month14[month])
    list=pd.read_csv(link)
    list.append(list)
    aprsep14=pd.DataFrame(list)

aprsep14.shape

split_datetime=pd.DataFrame(aprsep14['Date/Time'].str.split(' ').tolist(),
                         columns = ['Date','Time'])

data14=pd.concat([aprsep14,split_datetime],axis=1)
data14=data14.round(2)

## Converger les resultats avec Taxi zones pour obtenir le détail de l'historique data aprsep14

taxi_zone_coordinates=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_taxizones.csv",index_col=[0])
taxi_zone_coordinates.head()

result14=data14.merge(taxi_zone_coordinates, on=['Lat','Lon'], how='left').dropna()

result14=result14.reset_index().drop(['index'],axis=1)

result14=result14.rename(columns={"LocationID": "locationID","zone":"Zone","borough":"Borough"})
result14.head()

mycolumns=['Date','Time','locationID','Base','Lon','Lat','Borough','Zone']
result14=result14[mycolumns]
result14.head()

file_name = '/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_aprsep14.csv'
result14.to_csv(file_name)
print('File is written to CSV File successfully.')
