==> Uber Pickups/uber-trip-data/step3_aggregate_aprsep14_janjune15.ipynb <==
import pandas as pd
result14=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_aprsep14.csv",index_col=[0])
result15=pd.read_csv("/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/formatting_janjune15.csv",index_col=[0])

result14.head()

result14['Lat'].isnull().value_counts()

result15.head()

result15['Lat'].isnull().value_counts()

result1415=pd.concat([result14,result15],axis=0)

result1415

result1415['Lat'].isnull().value_counts()

file_name = '/Users/fezzibasma/Desktop/Data Full-Stack/Machine Learning Unsupervised/Projet Uber/export_step/aggregate_1415.csv'
result1415.to_csv(file_name)
print('File is written to CSV File successfully.')
